xcl2-python/requirements.txt
django~=5.2
djangorestframework==3.16.1
requests
PyJWT==2.8.0
python-decouple==3.8
drf-spectacular==0.27.2

